+++ src/main/java/com/example/budongbudong/domain/auction/dto/request/CreateAuctionRequest.java
@@ -4,6 +4,7 @@ import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Positive;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -14,7 +15,7 @@ public class CreateAuctionRequest {
 
     @NotNull(message = "시작가는 필수입니다.")
     @Positive(message = "시작가는 0보다 커야 합니다.")
-    private Long startPrice;
+    private BigDecimal startPrice;
 
     @NotNull(message = "경매 시작 시간은 필수입니다.")
     private LocalDateTime startedAt;
