+++ src/main/java/com/example/budongbudong/domain/property/document/PropertySearchDocument.java
@@ -3,10 +3,13 @@ package com.example.budongbudong.domain.property.document;
 import com.example.budongbudong.domain.property.enums.PropertyType;
 import lombok.*;
 import org.springframework.data.annotation.Id;
-import org.springframework.data.elasticsearch.annotations.*;
+import org.springframework.data.elasticsearch.annotations.Document;
+import org.springframework.data.elasticsearch.annotations.Field;
+import org.springframework.data.elasticsearch.annotations.FieldType;
+import org.springframework.data.elasticsearch.annotations.Setting;
 
 import java.math.BigDecimal;
-import java.time.*;
+import java.time.LocalDate;
 
 /**
  * Property 검색 전용 Elasticsearch Document
@@ -56,7 +59,7 @@ public class PropertySearchDocument {
 
     // --- 검색/정렬 ---
     @Field(type = FieldType.Long)
-    private Long price;
+    private BigDecimal price;
 
     // --- 경매 요약 (리스트용) ---
     @Field(type = FieldType.Object)
